Fix argument index
We hadn't updated the index we were using to retrieve the address
to interact with from the `args` slice. As `strconv.ParseUint()`
returns `0` in case of errors, we're always retrieving the data
from the first address, which was the one we're using when running
a local server... Anyway, good thing we noticed this now!